namada_tx_env
Trait TxEnv

pub trait TxEnv: StorageRead + StorageWrite {
    // Required methods
    fn read_bytes_temp(&self, key: &Key) -> Result<Option<Vec<u8>>>;
    fn write_bytes_temp(
        &mut self,
        key: &Key,
        val: impl AsRef<[u8]>,
    ) -> Result<()>;
    fn insert_verifier(&mut self, addr: &Address) -> Result<()>;
    fn init_account(
        &mut self,
        code_hash: impl AsRef<[u8]>,
        code_tag: &Option<String>,
        entropy_source: &[u8],
    ) -> Result<Address>;
    fn update_validity_predicate(
        &mut self,
        addr: &Address,
        code: impl AsRef<[u8]>,
        code_tag: &Option<String>,
    ) -> Result<()>;
    fn emit_event<E: EventToEmit>(&mut self, event: E) -> Result<()>;
    fn charge_gas(&mut self, used_gas: u64) -> Result<()>;
    fn get_events(&self, event_type: &EventType) -> Result<Vec<Event>>;
    fn set_commitment_sentinel(&mut self);
    fn update_masp_note_commitment_tree(
        transaction: &MaspTransaction,
    ) -> Result<bool>;

    // Provided methods
    fn read_temp<T: BorshDeserialize>(&self, key: &Key) -> Result<Option<T>> { ... }
    fn write_temp<T: BorshSerialize>(&mut self, key: &Key, val: T) -> Result<()> { ... }
}

Transaction host functions

Required Methods

fn read_bytes_temp(&self, key: &Key) -> Result<Option<Vec<u8>>>

Storage read temporary state raw bytes (after tx execution). It will try to read from only the write log.

fn write_bytes_temp(&mut self, key: &Key, val: impl AsRef<[u8]>) -> Result<()>

Write a temporary value as bytes at the given key to storage.

fn insert_verifier(&mut self, addr: &Address) -> Result<()>

Insert a verifier address. This address must exist on chain, otherwise the transaction will be rejected.

Validity predicates of each verifier addresses inserted in the transaction will validate the transaction and will receive all the changed storage keys and initialized accounts in their inputs.

fn init_account( &mut self, code_hash: impl AsRef<[u8]>, code_tag: &Option<String>, entropy_source: &[u8], ) -> Result<Address>

Initialize a new account generates a new established address and writes the given code as its validity predicate into the storage.

fn update_validity_predicate( &mut self, addr: &Address, code: impl AsRef<[u8]>, code_tag: &Option<String>, ) -> Result<()>

Update a validity predicate

fn emit_event<E: EventToEmit>(&mut self, event: E) -> Result<()>

Emit an Event from a transaction.

fn charge_gas(&mut self, used_gas: u64) -> Result<()>

Request to charge the provided amount of gas for the current transaction

fn get_events(&self, event_type: &EventType) -> Result<Vec<Event>>

Get events with a given EventType.

fn set_commitment_sentinel(&mut self)

Set the sentinel for an invalid section commitment

fn update_masp_note_commitment_tree( transaction: &MaspTransaction, ) -> Result<bool>

Update the masp note commitment tree in storage with the new notes

Provided Methods

fn read_temp<T: BorshDeserialize>(&self, key: &Key) -> Result<Option<T>>

Storage read temporary state Borsh encoded value (after tx execution). It will try to read from only the write log and then decode it if found.

fn write_temp<T: BorshSerialize>(&mut self, key: &Key, val: T) -> Result<()>

Write a temporary value to be encoded with Borsh at the given key to storage.

Object Safety

This trait is not object safe.