Crate namada_tx_env

Transaction environment contains functions that can be called from inside a tx.

Modules

storage

Storage types

Structs

Event

Custom events that can be queried from Tendermint using a websocket client

EventType

ABCI event type.

MaspTransaction

A MASP transaction.

Enums

Address

An account’s address

Traits

BorshDeserialize

A data-structure that can be de-serialized from binary format by NBOR.

BorshSerialize

A data-structure that can be serialized into binary format by NBOR.

BorshSerializeExt

Extensions to types implementing BorshSerialize.

EventToEmit

An event to be emitted in Namada.

ResultExt

Result extension to easily wrap custom errors into Error.

StorageRead

Common storage read interface

StorageWrite

Common storage write interface

TxEnv

Transaction host functions

Type Aliases

Result

Result of a storage API call.

Derive Macros

BorshDeserialize

BorshSerialize